fix(protocol-designer): fix `FlowRateField` max flow rate (#18165)
Previously, the range displayed in `FlowRateField` was pulled directly
from the pipette liquid definition. This value was specific only to the
pipette + tip combination. In reality, this max flow rate is a function
of the target volume, and correction volume for that
aspiration/dispense, as well as the pipette + tip accuracy function and
max plunger speed. This PR mirrors the hardware controller's logic for
determining the specific max allowable flow rate for the step.

Closes AUTH-1727