fix(api): fix SimulatedProbeResult ser/deser (#18177)
We weren't deserializing the simulation standin properly, which broke
runlog downloading. Now we are.

Also, we shouldn't be dumping it to the runlog anymore. That was
happening because of https://github.com/pydantic/pydantic/issues/6830
which causes mis-serialization of fields typed as `Model | NativeType`,
where `NativeType` is some scalar python type (i.e. float, int) and
`Model` is some pydantic model. With the model first, serializing the
container will always result in the serialization of the default
construction of the model, even if the attribute actually contained the
native type. Flipping the order makes it work correctly, for some
reason. Luckily this is consistent enough to be testable.

Closes RQA-4147
Closes EXEC-1429
Closes EXEC-1358