fix(app): fix labware displayed on modules during LPC (#18204)
Closes RQA-4156

When LPCing a labware in a module slot, the "Prepare deck for LPC" view's deck map currently renders the actively LPC'd labware above all modules and not just the module of interest.

To fix, we only render the LPC'd labware above the module with an id that matches the location-specific module id.