fix(api): handle stacks in move_labware
We altered modules to not be eventually loaded on deck slots on the
Flex, so we could better support the stacker providing two addressable
areas per module. This had an unexpected knock-on effect of making
get_ancestor_slot() not a good fallback if we ran out of depth to check
in get_labware_grip_point(), and protocols that called that
function (e.g. by calling move_labware) on a location that consisted of
a stack of more than one labware on a module (so, moving something on to
or off of a labware on an adapter on a module) would give you a deck
configuration conflict, since get_ancestor_slot would request a slot but
the module is there instead.

There is a larger fix for this - geometry needs to be more broadly
overhauled to work on location sequences throughout - but for now we can
make duals of the recursive pointer-chasing symbolic geometry resolvers
that work based on the "everything ends up on an addressable area"
assumption.

Particularly notable is modules - there was some very weird stuff in
here to make code that was intended for the OT-2 and geometry from the
module defs continue to work on Flex and geometry that is based on the
deck def addressable area coordinates, which was causing a lot of
complexity in get_labware_grip_point(). By changing the module
get_nominal_position() (or really making a new one) that returns the
offset of a child labware _from the nearest addressable area_ (i.e.
0-vec on the flex) instead of from the underlying deck slot, we can
eliminate a whole lot of logic.

Closes RQA-4153