refactor: unseal: 8chan: less z movement
Once we're moving DIRECT because we're actually saving our location, we
can just... not go as high. Rather than initially moving to the tip drop
location (which is basically guaranteed to be TOP), we can go and make
sure that (a) it's not TOP and (b) if it's defaulted then we can add an
offset to make it way lower, rather than going up high and then coming
back down.

This may not work at all for the 96.