fix(api): Do not hang if there is an error in `pickUpTip` planning (#18207)
## Overview

This closes RQA-4058 (protocol analysis hanging if `pickUpTip` is given
a bad labware definition), and resolves some duplication we noticed a
little while ago
(https://github.com/Opentrons/opentrons/pull/16469#discussion_r1797230213).

## Details

When a `pickUpTip` command completes, `TipStore` marks the picked-up
tips as "used." Before this PR, that worked by `PickUpTipImplementation`
passing along just the single "target" tip, like A1; `TipStore` then
internally expanded that into the full list of actually-picked-up tips,
like A1+B1+⋯+H1, for an 8-channel pickup.

The choice to do the expansion inside `TipStore` caused two problems:

1. Error handling. If it happens inside `TipStore`, and something goes
wrong and raises an exception, then we've raised an exception from
inside an `opentrons.protocol_engine` *state update.* We are not in a
good position to handle that: it was an early architectural of
`opentrons.protocol_engine` that all the real work ought to happen in
command implementations, and that state updates ought to be dead-simple
and always successful.

(We can and should do something about that premise, but that's beyond
the scope of this PR.)

This is what caused RQA-4058.

2. Duplication. For `TipStore` to automatically expand the target well
into the full list of actually-picked-up tips, it needed to know details
about the pipette's current configuration. That was already part of
`PipetteStore`'s job;`TipStore` was duplicating it.

This PR tries to solve both of those by having `PickUpTipImplementation`
give `TipStore` a full precomputed list of tips.
`PickUpTipImplementation` produces that precomputed list with help from
new functions in `PipetteView` and `TipView`. Pipette concerns like
nozzle maps are expunged from `TipStore`. Overall, there's no new logic
here—just shifting things around and changing what's running where.