refactor(api): fewer z raises in resin tip actions (#18212)
Avoid some previous uncommanded raises in Z while executing resin tip
commands.

These were caused by some unnecessarily high motions and because the
seal command didn't set the pipette locations tate update, so the
subsequent motions would run as-if moving between labware, raising the z
stage up to clear labware in between.

There was also an issue that required a slight refactor in `drop_tip`.
Our drop-tip functionality tries to position the pipette such that the
tip is somewhat inside the target well. This is based on multipliers
that are specific to the pipette and the tips, since we don't support
custom tips - we get to tune them in for ours. Well, these aren't our
tips, and to the system they're not really tips at all: we forbid using
tips as labware, and that's what resin tips are, so they can't be tips.
This added up to making the drop-tip (or, unseal in this case)
functionality drop the tips from way above the labware and they'd go
everywhere.

By overriding the function that provides the default drop tip location,
we can make this all work properly, and since we don't override it in
any other circumstances, we won't break the non-resin-tip drop.

## review
- good enough?

## testing
- [x] on 8 channel and 96 channel make sure that a straightforward
seal-dispense-unseal routine doesn't move high above the labware

Closes EXEC-1467