fix(api): discount labware in system location/hopper location when getting the deck height (#18188)
This PR adds a new geometry method to determine if labware is a deck
obstacle and updates the logic for calculating the highest Z-point on
the deck.

While refactoring the tests, I discovered that the we're also using the
wrong z for modules. We should only consider the module height within
the robot's deck, and not the overall height of the module.