fix(api): dont check SYSTEM lid stacks z
This avoids a problem where the protocol_engine_lid_stack object can be
in SYSTEM_LOCATION and we were asking for its ancestor slot. Also avoids
a problem where asking for the ancestor slot of a SYSTEM_LOCATION would
give you a bad kind of exception instead of a nice kind of exception.

Closes RQA-4164