fix(api): dont check SYSTEM lid stacks z (#18241)
This avoids a problem where the protocol_engine_lid_stack object can be
in SYSTEM_LOCATION and we were asking for its ancestor slot. Also avoids
a problem where asking for the ancestor slot of a SYSTEM_LOCATION would
give you a bad kind of exception instead of a nice kind of exception.

Closes RQA-4164

## Reviews and TODOs
This issue would also be addressed by
https://github.com/Opentrons/opentrons/pull/18188 but that includes
stacker stuff that isn't in the release branch, so it does just a much
lighter version of that.

## testing
Check that the protocol in the ticket passes analysis