fix(app): Fix ODD intermittent indefinite "canceling run" modal state (#18254)
Closes RQA-4112

The useRunStatus hook has proven to be bug prone in the past (with the desktop app's running protocol page tab transitions), and no longer provides any necessary functionality now that the run status does not transition to "idle" between protocol commands during a run.

Most likely, this hook was causing intermittent issues. There's also a chance we'll need to address EXEC-1450 to solve this entirely, but it's hard to say because it's so intermittent. If the run fetch happens to fail, let's try routing. Let's see if these changes solve the issue.

Given 8.4 release constraints, I've decided not to refactor out all the usages of useRunStatus, but I have created a sustaining ticket for that work at EXEC-1478. In the meantime, the hook has been marked as deprecated.