feat(opentrons-ai-client): UI feature for RTP if python format is selected (#18235)
# Overview

Closes AUTH-1662

This PR add `Runtime Parameters` section for UI. It is only available
for Flex.


![image](https://github.com/user-attachments/assets/ec810c29-c1f6-4330-ac71-c750c735b801)

So when robot is OT-2 or Protocol designer, it shows the message inside 


![image](https://github.com/user-attachments/assets/75336341-67cf-442c-8f9c-442df2b842b2)


Note that we don't want to completely remove the RTP section since all
sections are numbered, and removing it might affect references elsewhere
in the document. Therefore, this PR implements a straightforward
solution to address the issue while preserving the section structure.

## Test Plan and Hands on Testing

- Run and see the section is there.
- When you choose OT-2 or Protocol Designer, we are not able to enter
anything



## Risk assessment
Medium

<!--
- Indicate the level of attention this PR needs.
- Provide context to guide reviewers.
- Discuss trade-offs, coupling, and side effects.
- Look for the possibility, even if you think it's small, that your
change may affect some other part of the system.
- For instance, changing return tip behavior may also change the
behavior of labware calibration.
- How do your unit tests and on hands on testing mitigate this PR's
risks and the risk of future regressions?
- Especially in high risk PRs, explain how you know your testing is
enough.
-->