fix(app): move to initial calibration position after cancelling detach probe during desktop lpc (#18300)
Closes RQA-4177

On specifically the desktop app, it's possible to exit LPC on any view. During the detach view, we move the pipette to a maintenance position. If a user decides to cancel the detachment, we home the gantry and then return to whatever view the user previously visited.

This is all well and good except in one particular case: while actively calibrating an offset. We should move the pipette back to the initial position instead of the home position. This isn't currently a breaking bug, because we have the "start over" button on the desktop app as a workaround, but doing the correct movement implicitly is much nicer UX.

Although this seems like a decent amount of code, it's pretty straightforward: on specifically this calibrate offset view, we do the special homing behavior. We need a new command for this as well, because we don't want to reuse the existing command for moving to the initial position, since that one also does some labware loading.