feat(api): add option to air_gap() in place (#18249)
# Overview

The API `air_gap()` function allows you to specify a `height` above the
well to perform the air gap. If no `height` is specified, then
`air_gap()` defaults to 5 mm above the well. But that means there is no
way to air gap in place without moving the pipette.

To match the capabilities of PD step generation, this PR adds a new
`air_gap(in_place=True)` option to the API. This new parameter is gated
on API version 2.24 and above. AUTH-1797

## Test Plan and Hands on Testing

Added unit tests for the new `in_place` option, and made sure you cannot
specify both `in_place` and a `height`.

## Risk assessment

This is an API change, so once we release it publicly, we'll be stuck
with it.

(cherry picked from commit b13e976b46870a06237d1b4ce1a8aad490b810b8)