feat(step-generation): emit Python for `airGapInPlace` CommandCreator (#18314)
# Overview

After the recent PAPI changes to support `air_gap()` in-place, we can
now generate Python for the `airGapInPlace` CommandCreator.

This in turn completes the support for the `airGapInTrash` and
`airGapInWasteChute` compound commands.

Note that in the `airGapInTrash` and `airGapInWasteChute`
CommandCreators, we suppress the Python for `prepareToAspirate`, because
the Python `air_gap()` command automatically issues
`prepare_to_aspirate()` (PR #17919).

## Test Plan and Hands on Testing

Added unit tests.

I also tried PD Python export for a protocol that used air gaps, and
confirmed that it passed analysis.

## Risk assessment

Low.