feat(step-generation): emit Python for liquidProbe CommandCreator (#18322)
# Overview

We should emit Python for the `liquidProbe` CommandCreator to match the
behavior of JSON step generation. This is a follow-up to PR #18132.

I chose the Python `measure_liquid_height()` by trial-and-error. There
are 4 ways we can generate liquid probes in Python:
- `load_instrument(liquid_presence_detection=True)`
- `require_liquid_presence()`
- `detect_liquid_presence()`
- `measure_liquid_height()`

(1) `load_instrument(liquid_presence_detection=True)` adds `liquidProbe`
commands automatically for us before each `aspirate`. But that doesn't
match the command sequence of liquid class transfers, which adds
`liquidProbe` before `prepareToAspirate`.

(2) `require_liquid_presence()`, like its name implies, makes the
protocol fail if no liquid is present. That is not the behavior of our
JSON protocol, so we can't use that.

(3) `detect_liquid_presence()` inserts a `tryLiquidProbe` engine command
instead of `liquidProbe`. And for some reason, it probes from `z=0`
instead of from the `z=2` that we use in JSON step generation, so we
can't use that.

(4) So by process of elimination, that leaves `measure_liquid_height()`
as the only Python function that matches the JSON command.

I'm still not sure why we have so many different functions in Python,
and I'm not totally sure that we're doing the right thing with my choice
of `measure_liquid_height()`.

Stepping back, I'm also not sure if unconditionally adding
`liquidProbe`s to liquid class transfers is the right thing to do. But
this PR just focuses on generating Python code to match the JSON
commands.

## Test Plan and Hands on Testing

Added unit tests.
Looked at generated Python code manually.
Made sure generated Python protocols pass analysis.

## Risk assessment

Low.