fix(api): generate loc seqs for legacy (#18337)
We were not generating location sequences for the fake PE commands we
generate for the pre-2.14 pre-protocol-engine protocols, and since the
app is expecting them it wouldn't show any labware. Now we generate
them, because the comment about how it would be hard is just wrong (?)
and everything works.

Closes RQA-4181

## testing
- [x] put this on a robot
- [x] upload a protocol that is pre-2.14
- [x] pre-protocol should still show the labware

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: sfoster1 <3091648+sfoster1@users.noreply.github.com>
(cherry picked from commit df9ee0db60433245d0d26841856d1188daa35350)