fix(app): Hide the option to reset labware offsets on OT-2s (#18348)
## Overview

Closes RQA-4183.

In the device reset slideout, the "Clear labware offset data" option is
misleading on an OT-2. Although it does work in a technical sense—it
sends a `DELETE /labwareOffsets` request and the server obeys it—it does
not do anything meaningful in a user-facing sense. On OT-2s, the
frontend does not use the `/labwareOffsets` endpoints to support Labware
Position Check; it instead only reuses offsets from prior runs, via the
`/runs` endpoints. So this option didn't actually clear anything, in
practice.

The solution is to just hide the option on OT-2s.

## Test Plan and Hands on Testing

In the desktop app:

* On an OT-2:
    * [x] The option is no longer present in the desktop app.
* [x] Submitting the reset does not send a `DELETE /labwareOffsets` HTTP
request.
* On a Flex:
    * [x] The option is still present in the desktop app.
* [x] Submitting the reset with the option selected actually resets
labware offsets.

## Review requests

None in particular, but [I've definitely made stupid mistakes here
before](https://github.com/Opentrons/opentrons/pull/18244), so
double-check my logic.

## Risk assessment

Low.