fix(step-generation): emit waitForTemperature when checkbox is checked (#18370)
@sfoster1 noticed a bug where if you select the heater-shaker form
timer, it does not `waitForTemperature` if a `setTemperature` is set.

Basically there are 3 options for a heater-shaker form:

1. you build a heater-shaker step with no timer and choose to not add a
pause to wait for temperature: a `waitForTemperature` **is not**
created. this is correct
2. you build a heater-shaker step with no timer and choose to add a
pause to wait for temperature: a `waitForTemperature` **is** created,
this is correct
3. you build a heater-shaker step with a timer and you aren't prompted
to add a pause because you already added a timer: this is where the bug
is that this PR is addressing. It should emit a `waitForTemperature`
followed by a `waitForDuration` but it currently only emits a
`waitForDuration`. this is not correct