Widen getLabwareDefinitionUri().
No reason this can't work on both schema 3 and schema 2.

Also, the old type annotation of ProtocolFile<{}>['labwareDefinitions'] seemed semantically wrong. In protocol files, the labwareDefinitions object is keyed with arbitrary IDs, but in this util function, we specifically expect the keys to be labware URIs.

Also, this util doesn't really seem to be doing anything, but let's shelve that.