feat(app): Don't crash when we encounter labware schema 3 (#18397)
## Overview

This closes RQA-3998 and takes a step towards EXEC-1278.

## Changelog

This basically does a mass search-and-replace of `LabwareDefinition2` ->
`LabwareDefinition2 | LabwareDefinition3`.

Much of the time, this has no behavioral effect.

However, a lot of places in the frontend do math on the labware's
`cornerOffsetFromSlot` and `dimensions`. Those are the two things that
have changed so far from schema 2 to 3. (The other big change is
"locating features", which hasn't happened yet.) All of that math needs
to be evaluated on a case-by-case basis and ported to the schema 3 way
of doing things, replacing `cornerOffsetFromSlot` and `dimensions` with
`extents`.

**For scope reasons, this PR does not do that thinking. This PR naively
shims the math, knowingly causing visual errors:**

The idea is to do just enough to allow us to import custom schema-3
labware, and to run protocols that use schema-3 labware. This makes
testing easier, especially backend testing.

The rendering will be fixed in follow-up work.

The projects affected in this way are these:

* app (both desktop and ODD)
* components
* shared-data

Whereas these projects are, for now, remaining pinned to schema 2, for
scope reasons:

* labware-library
* labware-creator
* step-generation
* Quick Transfer Protocols (because they're based on step-generation)
* protocol-designer