fix(api): improve Flex Stacker shuttle homing routine (#18409)
# Overview
Before the platform sensor implementation, we defensively homed the Flex
Stacker to prevent platform collisions. This involved always retracting
the X-axis first, which was annoying because it moved in the opposite
direction of its home position. Now that we have platform sensor data,
we can make fewer assumptions and home the stacker more intelligently.