test(api): delete protocol_api_old/test_context.py? (#18412)
# Overview

Is this a good idea??!?

`protocol_api_old/test_context.py` is an old test that monkeypatches
`aspirate()` and `dispense()` in a very particular way:

```
def fake_aspirate(
    vol: Optional[float] = None,
    loc: Optional[Location] = None,
    rate: Optional[int] = None,
)
```

These argument names, default values, and number of arguments do not
match the current API at all. That means that to make the test pass, we
have to call `aspirate()` and `dispense()` with positional arguments
(not named arguments), and `instrument_context.py` has to avoid calling
`aspirate()`/`dispense()` with any of our new arguments that are not
available in the fake (like `push_out` or `flow_rate`).

I thought we couldn't change the test because it's enforcing some
ancient API behavior that we have to support. But @jbleon95 and
@SyntaxColoring think the whole test might be redundant now, so maybe we
can just delete it?

## Review requests

I'd like feedback from anyone who knows what `protocol_api_old` is, and
what the purpose of `protocol_api_old/test_context.py` is.

## Risk assessment

A test-only change.