feat(step-generation): refine advanced arg behavior for distribute (#18430)
Here I refine the logic for advanced argument behavior at destination
wells for a distribute chunk (touch tip, air gap, blowout). I now take
into account the correct air gap in the tip depending on which
destination well in the distribution chunk is being accessed. If the
first well in the chunk, the air gap to be dispensed is the aspirate air
gap volume, whereas if not the first well in the chunk, the air gap to
be dispensed is the dispense air gap.