fix(app): fix LPC error when loading a module into C2 (#18480)
Closes RESC-461

Loading a module, aka the mag block, into slot C2 causes LPC to error out when setting a default offset. Product + Design are working on a longer term solution, but in the interim, we want to include something for 8.5.

Select a default offset location for LPC based on slots that we know do not contain a module. We preferentially select C2 if available.