fix(labware-creator): Fix the preview viewbox not updating when it needs to (#18476)
## Overview

This fixes a few intertwined issues with the preview at the bottom of
Labware Creator not zooming and panning to show the whole labware like
it's intended to, closing EXEC-1607. It also tries to simplify the code,
which helps with EXEC-1573.

## Changelog

* Fix: Find the SVG dimensions *after* we render the SVG, not *before.*

We're using `useLayoutEffect` to render the SVG, find its dimensions,
and rerender it with a `viewBox` based on those dimensions before the
browser repaints the screen. That's all good so far.

But we were finding the dimensions (`getBBox()`) *outside* the
`useLayoutEffect()`, which meant we were getting the dimensions of the
SVG as it existed before we started rendering, not the new SVG that we
just rendered. This basically caused our `viewBox` value to always lag
behind our actual SVG contents by one render.

* Fix: Update the viewbox if its x- or y-offset changes. Formerly, we
were only updating it if its width or height changed.

* Refactor: Fully separate the empty-state placeholder from the
populated-state preview so there's no chance of their dimensions
affecting each other.

* Refactor: Simplify the dummy `viewBox` value that's used in the first
render, before we've measured the SVG and computed an actual value.
Formerly, it fell back to the labware definition's `dimensions` and then
to standard slot dimensions. Now, we just use a hard-coded `0 0 0 0`
because it's completely arbitrary and gets instantaneously replaced by a
real measured value.

---------

Co-authored-by: Jethary Alcid <66035149+jerader@users.noreply.github.com>
Co-authored-by: Jamey Huffnagle <jamey.huffnagle@opentrons.com>