chore(shared-data): lock in command schema 13 for RS 8.5.0, bump other changes to 14.json (#18475)
# Overview

We just cut `chore_release-8.5.0`, which is a limited cherrypick
release. We need to lock command schema `13.json` to the version of the
file that we're releasing in RS 8.5.0, and move all other changes to the
new `14.json`.

Note that this change is going into `chore_release-pd-8.5.0`. Again, our
strategy is that changes should flow from `chore_release-pd-8.5.0 ->
edge`.

## Test Plan and Hands on Testing

Will rely on CI testing.

## Risk assessment

Medium.