fix(app): do not select a default location that contains a fixture (#18508)
Closes RQA-4243

It's insufficient to only check module locations for the presence of an object on the deck, because doing so doesn't consider fixtures. Let's make things simpler: if a module is known to be in C2, using the LPC-able labware, let's find the labware that doesn't have a module beneath it and use that one.