feat(api): Speed up instrument calibration (#18469)
<!--
Thanks for taking the time to open a Pull Request (PR)! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

GitHub provides robust markdown to format your PR. Links, diagrams,
pictures, and videos along with text formatting make it possible to
create a rich and informative PR. For more information on GitHub
markdown, see:


https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview

So previously we were limiting the gantry speed for the calibration
process. This was a remnant of when we were considering doing a x/y
capacitive sweep instead of a z probe. This however does not give us
anything and just slows down the process.

Additionally we probe all 4 edges to account for any error, however we
also probe quite a large bounds on each edge. To preserve the ability to
overcome the same level of positional error I made it so that the first
x and the first y edges still use the older, larger sweeping region with
the same number of strides, however once we have a good idea of where
the first ones are we don't have to sweep such a large area for the
second two, so I reduced the starting stride distance and the number of
probes required for this.

All of this together the time to calibrate the pipette changed from
~2:10 from hitting go until the time it tells you to take off the probe
to 1:45.

The gripper also has benefits going from about ~4:10  to ~3:25
<!--
Describe your PR at a high level. State acceptance criteria and how this
PR fits into other work. Link issues, PRs, and other relevant resources.
-->

## Test Plan and Hands on Testing

<!--
Describe your testing of the PR. Emphasize testing not reflected in the
code. Attach protocols, logs, screenshots and any other assets that
support your testing.
-->

## Changelog

<!--
List changes introduced by this PR considering future developers and the
end user. Give careful thought and clear documentation to breaking
changes.
-->

## Review requests

<!--
- What do you need from reviewers to feel confident this PR is ready to
merge?
- Ask questions.
-->

## Risk assessment

<!--
- Indicate the level of attention this PR needs.
- Provide context to guide reviewers.
- Discuss trade-offs, coupling, and side effects.
- Look for the possibility, even if you think it's small, that your
change may affect some other part of the system.
- For instance, changing return tip behavior may also change the
behavior of labware calibration.
- How do your unit tests and on hands on testing mitigate this PR's
risks and the risk of future regressions?
- Especially in high risk PRs, explain how you know your testing is
enough.
-->